accuracy fix

+++ main.py
@@ -194,7 +194,7 @@ for epoch in range(wandb.config["NUM_EPOCHS"]):
         del tape
         # Update metrics
         metrics["Accuracy"].update_state(
-            y_batch_train, logits
+            tf.cast(y_batch_train, tf.int32), logits
         )
         metrics["TotalLoss"].update_state(loss_value)
         metrics["Routing0Loss"].update_state(routing_0_loss)
