+++ utils/validation.py
@@ -93,7 +93,7 @@ def validation(model, dataset, name, epoch, config, metrics, global_step, inform
         #     metrics["Route1"][c].update_state(tf.math.round(r_1))
 
         loss_value = classification_loss + routing_0_loss + routing_1_loss
-        metrics["Accuracy"].update_state(y_batch, logits)
+        metrics["Accuracy"].update_state(tf.cast(y_batch, tf.int32), logits)
         metrics["TotalLoss"].update_state(loss_value)
         metrics["Routing0Loss"].update_state(routing_0_loss)
         metrics["Routing1Loss"].update_state(routing_1_loss)
